requirements.txt
amqp==2.1.4
api==1.0.0.dev5
appdirs==1.4.3
beautifulsoup4==4.6.0
billiard==3.5.0.2
celery==4.0.2
cloudinary==1.8.0
coreapi==2.3.0
coreschema==0.0.4
defusedxml==0.5.0
Django==1.10.6
django-braces==1.11.0
django-choices==1.5.0
django-cors-headers==2.0.2
django-cors-middleware==1.3.1
django-filter==1.0.2
django-fsm==2.6.0
django-fsm-log==1.3.0
django-grappelli==2.9.1
django-mandrill==0.1.0
django-money==0.10.2
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.6
django-rest-swagger==2.1.2
djangorestframework==3.6.2
docopt==0.4.0
-e git+https://github.com/mobolic/facebook-sdk.git@e96186aa9244cae8e3a35d67d8cc8f94284a976e#egg=facebook_sdk
future==0.16.0
google-api-python-client==1.6.2
httplib2==0.10.3
itypes==1.1.0
Jinja2==2.9.6
kombu==4.0.2
lxml==3.8.0
mailchimp==2.0.9
mandrill==1.0.57
Markdown==2.6.8
MarkupSafe==1.0
mock==2.0.0
oauth2client==4.1.2
oauthlib==2.0.1
openapi-codec==1.3.1
packaging==16.8
pathlib==1.0.1
pbr==3.0.0
psycopg2==2.7.1
py-moneyed==0.7.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycryptodomex==3.4.6
pyjwkest==1.3.6
PyJWT==1.5.0
pyparsing==2.2.0
python-instagram==1.3.2
python-social-auth==0.3.6
python-twitter==3.3
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
seolib==0.1.3
simplejson==3.10.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
uritemplate==3.0.0
urllib3==1.21.1
vine==1.1.3
